backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
python-dotenv>=1.0
openai>=1.51
distro==1.9.0
langgraph>=0.2.40
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9